requirements.txt
# AI News Dashboard Requirements
# Core dependencies for the news dashboard and ML services

# Web Framework
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
redis==5.0.0

# Machine Learning & NLP
torch==2.0.1
torchvision==0.15.2
transformers==4.33.2
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
spacy==3.6.1
nltk==3.8.1

# Specific NLP Models
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl

# Deep Learning Frameworks
tensorflow==2.13.0
keras==2.13.1

# Data Processing
beautifulsoup4==4.12.2
requests==2.31.0
feedparser==6.0.10
newspaper3k==0.2.8

# Async & Concurrency
aiohttp==3.8.5
aioredis==2.0.1
celery==5.3.1

# Monitoring & Metrics
prometheus-client==0.17.1
grafana-api==1.0.3

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development Tools
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Utilities
dateutils==0.6.12
python-dateutil==2.8.2
tzlocal==5.0.1
geopy==2.3.0

# Security
cryptography==41.0.4
jwt==1.3.1
PyJWT==2.8.0

# Logging & Debugging
loguru==0.7.2
sentry-sdk==1.32.0

# API & HTTP
fastapi==0.103.1
uvicorn==0.23.2
httpx==0.24.1

# Image Processing (for visual news AI)
Pillow==10.0.0
opencv-python==4.8.0.76

# Scientific Computing
scipy==1.11.2
matplotlib==3.7.2
seaborn==0.12.2

# Caching & Performance
memcached==1.59
pymemcache==4.0.0

# Message Queues
kombu==5.3.2
billiard==4.1.0

# Deployment & DevOps
docker==6.1.3
kubernetes==27.2.0

# Optional: Advanced ML Libraries
# Uncomment if using advanced features
# xgboost==1.7.6
# lightgbm==4.0.0
# catboost==1.2
# optuna==3.3.0

# Optional: Quantum Computing (experimental)
# qiskit==0.44.1
# cirq==1.2.0

# Optional: Reinforcement Learning
# gym==0.29.1
# stable-baselines3==2.1.0

# Development Dependencies (install with pip install -r requirements-dev.txt)
# jupyter==1.0.0
# notebook==7.0.2
# ipykernel==6.25.1